refactor: Comment MVVM 구조 분리 + 코드 최적화
